frontend/requirements.txt
Flask==1.1.1
cryptography==2.8
python-jose[cryptography]==3.0.1
requests==2.22.0
google-cloud-storage==1.23.0
google-cloud-logging==1.14.0
google-cloud-error-reporting==0.33.0
google-cloud==0.34.0
google-cloud-datastore==1.12.0
